Rename folders using new naming convention
