+++ Amazon-Price-Alert/README.md
@@ -1,28 +1,34 @@
 # Amazon prices keep fluctuating- Let's scrape them!
 
- C3PO is a web-scraper built on BeautifulSoup that alerts you when the price of an amazon prduct falls within your budget! Currently in development.
- 
- ## What is the use?
- Nothing is worse than seeing the price drop of a product you just bought on Amazon drop 2 days after you bought it. Wouldn't it be amazing if a python script could just 
- ask you your budget and the product link and then notify you when the product price drops below that? This is exactly what this script does!
- 
- ## How to use 
- * Simply copy the link of the product and your budget that's it!
- * Paste the link when the program requests you to 
- * Enter budget for the product(without currency symbol) 
- * Select frequency to check prices
- * Keep the script running in the background
- The scraper will do the rest for you and notify you when the price is in your budget.
- However, it is to be kept in mind that the sender email ID and password has been stored in os variables in the script. Therefore, wherever you see the imports of 
- 'DEVELOPER_EMAIL' and 'DEVELOPER_PASS', be sure to replace them with the email ID you wish to recieve the notification from.
- ### Side Note
- Do remember to install the dependencies in the requirements.txt file!
- 
- ## Modules used (available in requirements.txt)
- * requests_html
- * BeautifulSoup
- 
- ## Development Status 
- This scrapper is complete. A future version may have emails sent via a server.
- 
-  ### Developed by [Sarthak Saxena](https://github.com/sarthak1905) 
+C3PO is a web-scraper built on BeautifulSoup that alerts you when the price of an amazon prduct falls within your budget! Currently in development.
+
+## What is the use?
+
+Nothing is worse than seeing the price drop of a product you just bought on Amazon drop 2 days after you bought it. Wouldn't it be amazing if a python script could just
+ask you your budget and the product link and then notify you when the product price drops below that? This is exactly what this script does!
+
+## How to use
+
+- Simply copy the link of the product and your budget that's it!
+- Paste the link when the program requests you to
+- Enter budget for the product(without currency symbol)
+- Select frequency to check prices
+- Keep the script running in the background
+  The scraper will do the rest for you and notify you when the price is in your budget.
+  However, it is to be kept in mind that the sender email ID and password has been stored in os variables in the script. Therefore, wherever you see the imports of
+  'DEVELOPER_EMAIL' and 'DEVELOPER_PASS', be sure to replace them with the email ID you wish to recieve the notification from.
+
+### Side Note
+
+Do remember to install the dependencies in the requirements.txt file!
+
+## Modules used (available in requirements.txt)
+
+- requests_html
+- BeautifulSoup
+
+## Development Status
+
+This scrapper is complete. A future version may have emails sent via a server.
+
+### Developed by [Sarthak Saxena](https://github.com/sarthak1905)
