+++ Binary Calculator [GUI]/script.py
