+++ Screenshot [GUI]/screenshot_gui.py
